refactor tips.*pp, add path_jaccard.*pp
